chore: change operation amount to bigint (#21)

+++ apps/banking_api_challenge/priv/repo/migrations/20210416144125_create_operations_table.exs
@@ -5,7 +5,7 @@ defmodule BankingApiChallenge.Repo.Migrations.CreateOperationsTable do
     create table(:operations, primary_key: false) do
       add :id, :uuid, primary_key: true
       add :operation_type, :string, null: false
-      add :amount, :integer, null: false, default: 0
+      add :amount, :bigint, null: false, default: 0
 
       add :source_account_id, references(:accounts, type: :uuid)
       add :target_account_id, references(:accounts, type: :uuid)
